In settingsController, Fiffanonymizer app list is stored in Heap.
We are working here with 100 logical core machines, thus in the foreseable future we might be looking at several 100s applications.
Right now the app would store all these in the stack. To avoid future problems, lets use the heap for what it was designed
for.
Besides, it is always a joy to use the term, for what it was originally purposed, instead of a very wellknown programming website. (Guess what webpage I'm thinking of?)... ;)